Fix Windows build

+++ common/JammerNetzPackage.cpp
@@ -313,7 +313,7 @@ std::shared_ptr<AudioBlock> JammerNetzAudioData::readAudioHeaderAndBytes(JammerN
 	result->sampleRate = 48000;
 	size_t upsampleRate = block->sampleRate() != 0 ? 48000 / block->sampleRate() : 48000;
 	jassert(block->numberOfSamples() * upsampleRate == SAMPLE_BUFFER_SIZE);
-	result->audioBuffer = std::make_shared<AudioBuffer<float>>(block->numChannels(), block->numberOfSamples() * upsampleRate);
+	result->audioBuffer = std::make_shared<AudioBuffer<float>>((int) block->numChannels(), (int) (block->numberOfSamples() * upsampleRate));
 	readAudioBytes(block->channels(), result->audioBuffer, upsampleRate);
 	return result;
 }
